backend/requirements.txt
# Flask and web framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests and API clients
requests==2.31.0
urllib3==2.0.7

# JSON and data handling
jsonschema==4.19.2

# Date and time handling
python-dateutil==2.8.2

# File handling and utilities
Pillow==10.0.1
pathlib2==2.3.7

# Logging and monitoring
colorlog==6.7.0

# Background tasks and scheduling
APScheduler==3.10.4
celery==5.3.4

# Database (if needed later)
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# Authentication and security
PyJWT==2.8.0
bcrypt==4.0.1

# Social media APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Instagram automation
selenium==4.15.2
webdriver-manager==4.0.1

# Video and image processing
opencv-python==4.8.1.78
moviepy==1.0.3
imageio==2.31.5

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# AI and ML utilities
numpy==1.24.4
scipy==1.11.4

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
black==23.10.1
flake8==6.1.0